Moved the ps-storing functions to a more a appropriate location
