+++ ExplanationGeneration/MinedPSManager.py
@@ -4,11 +4,10 @@ import numpy as np
 
 from BenchmarkProblems.BenchmarkProblem import BenchmarkProblem
 from FirstPaper import TerminationCriteria
-from FirstPaper.EvaluatedPS import EvaluatedPS
+from FirstPaper.EvaluatedPS import EvaluatedPS, write_evaluated_pss_to_file, load_pss
 from FirstPaper.PRef import PRef
-from FirstPaper.PS import PS, STAR
+from FirstPaper.PS import PS, STAR, write_pss_to_file
 from PSMiners.AbstractPSMiner import AbstractPSMiner
-from PSMiners.Mining import write_evaluated_pss_to_file, load_pss, write_pss_to_file
 from PSMiners.PyMoo.SequentialCrowdingMiner import SequentialCrowdingMiner
 from utils import announce
